refactor: Introduce AddActionButton component across various pages and components,
