+++ src/components/atoms/Button/index.ts
@@ -1,3 +1,4 @@
 export { default as Button } from './Button';
 export { default as AddButton } from './AddButton';
 export type { ButtonProps } from './Button';
+export { default as AddActionButton } from './AddActionButton';
