+++ src/components/atoms/index.ts
@@ -1,4 +1,4 @@
-export { Button, AddButton } from './Button';
+export { Button, AddButton, AddActionButton } from './Button';
 export { default as Input } from './Input';
 export { default as SectionHeader } from './SectionHeader';
 export { default as Chip } from './Chip';
