+++ src/components/organisms/PlanForm/index.ts
@@ -1,3 +1,2 @@
 export { default as PlanDetailsSection } from './PlanDetailsSection';
-export { default as SetupChargesSection } from './SetupChargesSection';
 export { default as RecurringChargesForm } from './RecurringChargesForm';
